# apps/dashboard_app/src/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/version":{"get":{"summary":"Get Backend Version","operationId":"get_backend_version_api_v1_version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Get Backend Version Api V1 Version Get"}}}}}}},"/api/v1/action-cards":{"get":{"summary":"List Action Cards","operationId":"list_action_cards_api_v1_action_cards_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionCards"}}}}}}},"/api/v1/session":{"get":{"summary":"Get Session","operationId":"get_session_api_v1_session_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkshopSessionResponse"}}}}}},"put":{"summary":"Update Session","operationId":"update_session_api_v1_session_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkshopSessionUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkshopSessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Session","operationId":"create_session_api_v1_session_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkshopSessionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkshopSessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scenario/action-cards":{"get":{"summary":"List Scenario Action Cards","operationId":"list_scenario_action_cards_api_v1_scenario_action_cards_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionCards"}}}}}}},"/api/v1/scenario/scores":{"get":{"summary":"Get Scenario Scores","operationId":"get_scenario_scores_api_v1_scenario_scores_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScenarioScoresResponse"}}}}}}},"/api/v1/teams":{"get":{"summary":"List Groups","operationId":"list_groups_api_v1_teams_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamsResponse"}}}}}},"post":{"summary":"Create Team","operationId":"create_team_api_v1_teams_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teams/evaluation":{"get":{"summary":"Get Teams Evaluation","operationId":"get_teams_evaluation_api_v1_teams_evaluation_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamsEvaluationResponse"}}}}}}},"/api/v1/teams/{team_id}":{"get":{"summary":"Get Team","operationId":"get_team_api_v1_teams__team_id__get","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"string","title":"Team Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Team","operationId":"delete_team_api_v1_teams__team_id__delete","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"string","title":"Team Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teams/{team_id}/action-cards":{"get":{"summary":"List Team Action Cards","operationId":"list_team_action_cards_api_v1_teams__team_id__action_cards_get","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"string","title":"Team Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionCards"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Add Team Action Cards","operationId":"add_team_action_cards_api_v1_teams__team_id__action_cards_post","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"string","title":"Team Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionCardsCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionCards"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Team Action Cards","operationId":"delete_team_action_cards_api_v1_teams__team_id__action_cards_delete","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"string","title":"Team Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionCardsCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionCards"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ActionCard":{"properties":{"card_id":{"type":"string","title":"Card Id"},"scenario":{"type":"string","title":"Scenario"},"card_number":{"type":"integer","title":"Card Number"},"cost":{"type":"number","title":"Cost"},"time":{"type":"number","title":"Time"},"efficiency":{"type":"number","title":"Efficiency"}},"type":"object","required":["card_id","scenario","card_number","cost","time","efficiency"],"title":"ActionCard"},"ActionCardCreate":{"properties":{"card_id":{"type":"string","title":"Card Id"}},"type":"object","required":["card_id"],"title":"ActionCardCreate"},"ActionCards":{"properties":{"action_cards":{"items":{"$ref":"#/components/schemas/ActionCard"},"type":"array","title":"Action Cards"}},"type":"object","required":["action_cards"],"title":"ActionCards"},"ActionCardsCreate":{"properties":{"action_cards":{"items":{"$ref":"#/components/schemas/ActionCardCreate"},"type":"array","title":"Action Cards"}},"type":"object","required":["action_cards"],"title":"ActionCardsCreate"},"BestScores":{"properties":{"time":{"type":"integer","title":"Time"},"cost":{"type":"integer","title":"Cost"},"efficiency":{"type":"integer","title":"Efficiency"}},"type":"object","required":["time","cost","efficiency"],"title":"BestScores"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MaximaScores":{"properties":{"time":{"type":"integer","title":"Time"},"cost":{"type":"integer","title":"Cost"},"efficiency":{"type":"integer","title":"Efficiency"}},"type":"object","required":["time","cost","efficiency"],"title":"MaximaScores"},"ScenarioScoresResponse":{"properties":{"maxima":{"$ref":"#/components/schemas/MaximaScores"},"best":{"$ref":"#/components/schemas/BestScores"}},"type":"object","required":["maxima","best"],"title":"ScenarioScoresResponse"},"Team":{"properties":{"name":{"type":"string","title":"Name"},"id":{"type":"string","title":"Id"},"has_all_cards":{"type":"boolean","title":"Has All Cards"},"action_cards":{"items":{"$ref":"#/components/schemas/ActionCard"},"type":"array","title":"Action Cards"}},"type":"object","required":["name","id","has_all_cards","action_cards"],"title":"Team"},"TeamCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"TeamCreate"},"TeamResponse":{"properties":{"team":{"$ref":"#/components/schemas/Team"}},"type":"object","required":["team"],"title":"TeamResponse"},"TeamsEvaluation":{"properties":{"time":{"items":{"type":"integer"},"type":"array","title":"Time"},"cost":{"items":{"type":"integer"},"type":"array","title":"Cost"},"efficiency":{"items":{"type":"integer"},"type":"array","title":"Efficiency"}},"type":"object","required":["time","cost","efficiency"],"title":"TeamsEvaluation"},"TeamsEvaluationResponse":{"properties":{"teams_evaluation":{"$ref":"#/components/schemas/TeamsEvaluation"},"totals_evaluation":{"$ref":"#/components/schemas/TotalsEvaluation"},"team_names":{"items":{"type":"string"},"type":"array","title":"Team Names"}},"type":"object","required":["teams_evaluation","totals_evaluation","team_names"],"title":"TeamsEvaluationResponse"},"TeamsResponse":{"properties":{"teams":{"items":{"$ref":"#/components/schemas/Team"},"type":"array","title":"Teams"}},"type":"object","required":["teams"],"title":"TeamsResponse"},"TotalsEvaluation":{"properties":{"time":{"type":"number","title":"Time"},"cost":{"type":"number","title":"Cost"},"efficiency":{"type":"number","title":"Efficiency"}},"type":"object","required":["time","cost","efficiency"],"title":"TotalsEvaluation"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WorkshopSession":{"properties":{"scenario_id":{"type":"string","title":"Scenario Id"},"state":{"$ref":"#/components/schemas/WorkshopState"},"id":{"type":"string","title":"Id"}},"type":"object","required":["scenario_id","state","id"],"title":"WorkshopSession"},"WorkshopSessionCreate":{"properties":{"scenario_id":{"type":"string","title":"Scenario Id"}},"type":"object","required":["scenario_id"],"title":"WorkshopSessionCreate"},"WorkshopSessionResponse":{"properties":{"workshop_session":{"anyOf":[{"$ref":"#/components/schemas/WorkshopSession"},{"type":"null"}]}},"type":"object","required":["workshop_session"],"title":"WorkshopSessionResponse"},"WorkshopSessionUpdate":{"properties":{"state":{"$ref":"#/components/schemas/WorkshopState"}},"type":"object","required":["state"],"title":"WorkshopSessionUpdate"},"WorkshopState":{"type":"string","enum":["INTRODUCTION","TEAM_NAMING","SCANNING","EVALUATION"],"title":"WorkshopState"}}}}